requirements.txt
python-gitlab==1.15.0
python-telegram-bot==12.2.0